implement the ability to return only the last packet present in buffer
For sensors, it might be better to return only the last packet
present in the buffer (i.e. the newest sample) instead of the first
one (the oldest). This commit implements this functionality.

Moreover, the overall packet extraction logic has been refactored
so that it is done at the same place for both the internal buffer
and for the external one. Given that this logic has got complicated,
it is better to have it in one place ;-)

Conflicts: