fix an obvious problem with the current implementation of the last_packet thing
This commit smoothes the buffer management by doing all the
work in the internal buffer, copying only relevant data into
the user-provided buffer. It means that the buffer is not
touched if no data has been found (which is always good).

The goal here was to fix a problem, where the last_packet flag
was not honored in all cases. The new implementation will now
really empty the kernel IO buffer before deciding whether or not
the last packet has been found.