add an optional behaviour to flush all pending data in the internal buffer on timeout
This is required in cases where it is expected that corrupted data
would come in (i.e. corrupting medium). Otherwise, the driver might get
stuck at a broken packet, trying forever for data.

The behaviour, when enabled, is to try and read packets from the
immediately available data, discarding everything that is not a full
packet -- it would for instance discard a partial packet.