Implement the MOCK funcionality to be able to send messages that require a reply while testing Drivers
The MOCK functionality must be set with IODRIVERS_BASE_MOCK, what puts
the driver in mock mode. The expectation tries to mimic the GTEST style
and are set with EXPECT_REPLY(expectation,reply).

Multiple expecatitons can be set and must be satisfied in the order of
creation. If any expecation is not met at the end of the test, the test
fails.