give the means to test openURI itself
The problem here was that openURI was the one creating the test stream,
so the fixture class could not pass data to it. The change is to allow
creation of the

Note that it also allows to test "lingering data" handling, which may
happen when the device sent data either before the device was opened
(on HW buffered lines like serial), or between the open and the first
read.

It also handles situations where some data was not read after the last
close.