refactor readRaw to write directly to the output buffer
The data flow, going through the internal buffer, was
misinterpreted on review - so, most likely not clear.

Refactor to avoid the intermediate, instead writing straight to the
output buffer. The way the deadline is handled also avoids having
the do { ... } while, which was making the timeout logic harder
to comprehend IMO.