refactor readPacket to make the whole timeout logic clearer
The Timeout class has difficult to understand semantics, and
re-evaluates now() for each call, which makes the elapsed/timeLeft
combination hard to understand: one has to know that timeLeft
returns 0 if the timeout has elapsed to realize that while there
is a race between elapsed and timeLeft, it only results in an
unnecessary read. The refactor does away with Timeout to instead
explicitly compute deadlines and reason about them. This results
to less state maintained in the loop logic.

Moreover, the special case of first_byte_timeout == -1 is removed.
The only user-visible impact is that exceptions that would before
have PACKET as timeout type now have FIRST_BYTE if no bytes were
received, something that I would argue is an improvement. The overall
deadline is not changed, as it is computed from the start time (as
it was already).